Fix logic-to-bit assignment in 4-state e2e test
Cast via `as u8` instead of bare assignment, matching the Rust-side
integration test and satisfying the Veryl analyzer's type check.

+++ packages/celox/src/e2e.test.ts
@@ -445,7 +445,7 @@ module LogicBitMix (
     y_logic_from_bit: output logic<8>,
     y_bit_from_logic: output bit<8>,
 ) {
-    assign y_bit_from_logic = a_logic;
+    assign y_bit_from_logic = a_logic as u8;
     assign y_logic_from_bit = b_bit;
 }
 `;
